Move ConfigClient to client module
