+++ src/daq_config_server/__init__.py
@@ -7,6 +7,10 @@
 """
 
 from ._version import __version__
-from .app.client import ConfigClient
 
-__all__ = ["__version__", "ConfigClient"]
+# from .client._client import ConfigClient
+
+__all__ = [
+    "__version__",
+    # "ConfigClient"
+]
