+++ src/daq_config_server/app/__init__.py
@@ -1,4 +1,3 @@
-from ._server_response import MockPathToConverterDict
 from .api import main
 
-__all__ = ["MockPathToConverterDict", "main"]
+__all__ = ["main"]
