+++ src/daq_config_server/client/__init__.py
@@ -1,0 +1,3 @@
+from ._client import ConfigClient
+
+__all__ = ["ConfigClient"]
