+++ src/daq_config_server/client/_server_response.py
@@ -8,9 +8,8 @@ import requests
 from requests import Response as RealResponse
 from requests.exceptions import HTTPError
 
-from daq_config_server.models import ConfigModel
-
-from ._routes import ValidAcceptHeaders
+from daq_config_server.app.endpoints import ValidAcceptHeaders
+from daq_config_server.models.base_model import ConfigModel
 
 NonModel = str | bytes | dict[str, Any]
 MockPathToConverterDict = dict[Path, Callable[[str], ConfigModel | NonModel]]
